requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
dateparser==1.1.8
greenlet==2.0.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.7
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.14
pytz==2023.3
regex==2023.6.3
requests==2.31.0
rich==13.4.2
schedule==1.2.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tornado==6.1
typing-extensions==4.7.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
